requirements.txt
google-generativeai
requests
python-dotenv
rich
pydantic
pytest